Fix a bug in timer set

+++ src/codegen/util/bloom_filter.cpp
@@ -55,8 +55,8 @@ void BloomFilter::Init(uint64_t estimated_num_tuples) {
                 log(1 - pow(kFalsePositiveRate, 1.0 / (double)num_hash_funcs_));
   }
   LOG_INFO("BloomFilter num_bits: %lu bits_per_element: %f num_hash_funcs: %lu",
-           num_bits_, (double)num_bits_ / estimated_num_tuples,
-           num_hash_funcs_);
+           (unsigned long)num_bits_, (double)num_bits_ / estimated_num_tuples,
+           (unsigned long)num_hash_funcs_);
 
   // Allocate memory for the underlying bytes array
   uint64_t num_bytes = (num_bits_ + 7) / 8;
@@ -71,7 +71,7 @@ void BloomFilter::Init(uint64_t estimated_num_tuples) {
 void BloomFilter::Destroy() {
   // Free memory of underlying bytes array
   LOG_DEBUG("Bloom Filter, num_probes: %lu, misses: %lu, Selectivity: %f",
-            num_probes_, num_misses_,
+            (unsigned long)num_probes_, (unsigned long)num_misses_,
             (double)(num_probes_ - num_misses_) / num_probes_);
   delete[] bytes_;
 }
